feat(project): move pwned matcher into own package
