+++ packages/libraries/main/src/index.ts
@@ -3,7 +3,6 @@ import scoring from './scoring'
 import TimeEstimates from './TimeEstimates'
 import Feedback from './Feedback'
 import Options from './Options'
-import matcherPwnedFactory from './matcher/pwned'
 import { MatchEstimated, MatchExtended, ZxcvbnResult } from './types'
 
 const time = () => new Date().getTime()
@@ -56,5 +55,3 @@ export const zxcvbn = (
 }
 
 export { Options as ZxcvbnOptions, ZxcvbnResult }
-
-export { matcherPwnedFactory as ZxcvbnMatcherPwnedFactory }
