+++ packages/libraries/main/tsconfig.json
@@ -1,7 +1,3 @@
 {
-  "extends": "../../../tsconfig.json",
-  "compilerOptions": {
-    "lib": ["es2018", "dom"],
-    "types": ["node", "jest"]
-  }
+  "extends": "../../../tsconfig.json"
 }
