+++ packages/libraries/pwned/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "@zxcvbn-ts/matcher-pwned",
+  "version": "1.0.0",
+  "main": "dist/index.js",
+  "module": "dist/index.esm.js",
+  "types": "dist/index.d.ts",
+  "description": "HaveIBeenPwned Matcher for zxcvbn-ts",
+  "license": "MIT",
+  "homepage": "https://github.com/zxcvbn-ts/zxcvbn",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/zxcvbn-ts/zxcvbn.git"
+  },
+  "author": "zxcvbn-ts",
+  "files": [
+    "dist",
+    "src"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "peerDependencies": {
+    "@zxcvbn-ts/core": "^1.0.0"
+  },
+  "keywords": [
+    "pwned",
+    "password",
+    "passphrase",
+    "security",
+    "quality",
+    "estimation",
+    "pattern",
+    "cracking",
+    "scoring",
+    "haveibeenpwned",
+    "zxcvbn-ts"
+  ]
+}
