+++ packages/libraries/pwned/src/scoring.ts
@@ -1,0 +1,4 @@
+// TODO make some more appropriated guesses logic?
+export default () => {
+  return 1
+}
