+++ packages/libraries/pwned/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../../tsconfig.json",
+  "compilerOptions": {
+    "lib": ["es2018", "dom"],
+    "types": ["node", "jest"]
+  }
+}
